[ch#160822794]Structure the db files to fit the project
