+++ server/app/__init__.py
@@ -6,3 +6,4 @@ def create_app(Config_Status):
     app.config.from_object(App_Configurations[Config_Status])
 
     return app
+
